refactor(push): extract notification service helper functions
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
